Update Farkle.py
list_melds now returns detailed information about the roll as a nested dictionary.
Completed algorithm for player meld choice entry and input validation with error tracking.
Next is to make the choice to continue the player turn with remaining dice or to end player turn and bank points gained this turn.